Cloud Formation Template used to install WordPress using Chef Repo CookBook
Cloud Formation Template that is designed to install WordPress using Chef + ChefDK. A Chef cookbook is downloaded from the Chef Cookbook site. The Berksfile is used to create Berks-Cookbooks that are then used to be installed on the local server. Tested the script, it does work, but there is a dependency issue with 7-zip due to Nginx issue.